+++ apps/cow-fi/CHANGELOG.md
@@ -1,5 +1,60 @@
 # Changelog
 
+## [2.5.8](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.5.7...cow-fi-v2.5.8) (2026-05-12)
+
+
+### 🐛 Bug Fixes
+
+* resolve transitive deps ([57d4b0f](https://github.com/cowprotocol/cowswap/commit/57d4b0f3236bdec7d3cd489fd663557ea31e6062))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 3.2.2
+    * @cowprotocol/assets bumped to 2.3.0
+    * @cowprotocol/common-hooks bumped to 3.2.2
+    * @cowprotocol/common-utils bumped to 3.3.2
+    * @cowprotocol/core bumped to 3.2.2
+    * @cowprotocol/ui bumped to 3.4.0
+    * @cowprotocol/wallet bumped to 3.3.0
+
+## [2.5.7](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.5.6...cow-fi-v2.5.7) (2026-05-08)
+
+
+### 🐛 Bug Fixes
+
+* bump next version ([#7471](https://github.com/cowprotocol/cowswap/issues/7471)) ([5f0d60f](https://github.com/cowprotocol/cowswap/commit/5f0d60f1933ca435e49138c8c2133341bfd5bf9e))
+
+## [2.5.6](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.5.5...cow-fi-v2.5.6) (2026-04-30)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/widget-react bumped to 3.0.2
+
+## [2.5.5](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.5.4...cow-fi-v2.5.5) (2026-04-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 3.2.1
+    * @cowprotocol/common-const bumped to 3.2.1
+    * @cowprotocol/common-hooks bumped to 3.2.1
+    * @cowprotocol/common-utils bumped to 3.3.1
+    * @cowprotocol/core bumped to 3.2.1
+    * @cowprotocol/events bumped to 4.3.1
+    * @cowprotocol/types bumped to 4.3.1
+    * @cowprotocol/ui bumped to 3.3.1
+    * @cowprotocol/wallet bumped to 3.2.1
+    * @cowprotocol/wallet-provider bumped to 2.1.17
+    * @cowprotocol/widget-react bumped to 3.0.1
+
 ## [2.5.4](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.5.3...cow-fi-v2.5.4) (2026-04-16)