+++ apps/widget-configurator/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Changelog
 
+## [3.4.3](https://github.com/cowprotocol/cowswap/compare/widget-configurator-v3.4.2...widget-configurator-v3.4.3) (2026-05-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 3.2.2
+    * @cowprotocol/assets bumped to 2.3.0
+    * @cowprotocol/common-hooks bumped to 3.2.2
+    * @cowprotocol/common-utils bumped to 3.3.2
+
+## [3.4.2](https://github.com/cowprotocol/cowswap/compare/widget-configurator-v3.4.1...widget-configurator-v3.4.2) (2026-04-30)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/widget-lib bumped to 4.0.2
+    * @cowprotocol/widget-react bumped to 3.0.2
+
+## [3.4.1](https://github.com/cowprotocol/cowswap/compare/widget-configurator-v3.4.0...widget-configurator-v3.4.1) (2026-04-22)
+
+
+### 🐛 Bug Fixes
+
+* take protocol fee into account ([#7385](https://github.com/cowprotocol/cowswap/issues/7385)) ([53057f8](https://github.com/cowprotocol/cowswap/commit/53057f8becaa6cfe9182e2d726471794d03b5775))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 3.2.1
+    * @cowprotocol/common-const bumped to 3.2.1
+    * @cowprotocol/common-hooks bumped to 3.2.1
+    * @cowprotocol/common-utils bumped to 3.3.1
+    * @cowprotocol/events bumped to 4.3.1
+    * @cowprotocol/types bumped to 4.3.1
+    * @cowprotocol/widget-lib bumped to 4.0.1
+    * @cowprotocol/widget-react bumped to 3.0.1
+
 ## [3.4.0](https://github.com/cowprotocol/cowswap/compare/widget-configurator-v3.3.3...widget-configurator-v3.4.0) (2026-04-16)