+++ libs/common-hooks/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [3.2.2](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.2.1...common-hooks-v3.2.2) (2026-05-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-utils bumped to 3.3.2
+
+## [3.2.1](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.2.0...common-hooks-v3.2.1) (2026-04-22)
+
+
+### 🐛 Bug Fixes
+
+* take protocol fee into account ([#7385](https://github.com/cowprotocol/cowswap/issues/7385)) ([53057f8](https://github.com/cowprotocol/cowswap/commit/53057f8becaa6cfe9182e2d726471794d03b5775))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 3.2.1
+    * @cowprotocol/common-utils bumped to 3.3.1
+    * @cowprotocol/types bumped to 4.3.1
+
 ## [3.2.0](https://github.com/cowprotocol/cowswap/compare/common-hooks-v3.1.2...common-hooks-v3.2.0) (2026-04-16)